adding CRUD for address, region, locality, locality-detail and governorate

+++ Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/address/Address.java
@@ -24,7 +24,7 @@ public class Address implements Serializable {
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idAddress;
+    private Long idAddress;
     private String description;
     @OneToOne(mappedBy = "address")
     private RealEstateListing realEstateListing;
