+++ Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/governorate/GovernorateRepository.java
@@ -2,6 +2,6 @@ package tn.esprit.propnetapp.governorate;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GovernorateRepository extends JpaRepository<Governorate, Integer> {
+public interface GovernorateRepository extends JpaRepository<Governorate, Long> {
 
 }
