+++ Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/locality/LocalityRepository.java
@@ -2,6 +2,6 @@ package tn.esprit.propnetapp.locality;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalityRepository extends JpaRepository<Locality, Integer> {
+public interface LocalityRepository extends JpaRepository<Locality, Long> {
 
 }
