+++ Backend/PropNetApp/src/main/java/tn/esprit/propnetapp/localitydetail/LocalityDetailRepository.java
@@ -2,6 +2,6 @@ package tn.esprit.propnetapp.localitydetail;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalityDetailRepository extends JpaRepository<LocalityDetail, Integer> {
+public interface LocalityDetailRepository extends JpaRepository<LocalityDetail, Long> {
 
 }
